chhotu/requirements.txt
django
djangorestframework
drfdocs